Spawn and pathing for wolf chase
Added spawn point at botttom of a mountain for the wolf chase scene. I made some trees with fire as a starting point to guide the rabbit up the mountain. I added one falling tree but I diddnt finish implementing it. The tree pathing needs to be finished and fire walls also need to be added.